test(Playwright): Move Richtext Helpers to fixture/
`Richtext` provides test fixtures, so `fixture` seems to be a good
location to hold this utility.
